chore: code placement change of handler metric
